requirements.txt
flask
python-dotenv
pdfplumber
flask-cors
requests